requirements.txt
# local packages
-e .

#other dependencies
tqdm
dvc
pandas
numpy
SciPy
PyYAML
scikit-learn
lxml
botcore